Refactoring: Merge compiler and jit-rt optimizer.cpp versions
